+++ src/Model/broker/StocksInterface.py
@@ -1,0 +1,43 @@
+import os
+import sys
+import inspect
+import functools
+import threading
+
+currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
+parentdir = os.path.dirname(currentdir)
+sys.path.insert(0, parentdir)
+
+# Mutex used for thread synchronisation
+lock = threading.Lock()
+
+
+def synchronised(lock):
+    """ Thread synchronization decorator """
+
+    def wrapper(f):
+        @functools.wraps(f)
+        def inner_wrapper(*args, **kw):
+            with lock:
+                return f(*args, **kw)
+
+        return inner_wrapper
+
+    return wrapper
+
+
+class Singleton(type):
+    """Metaclass to implement the Singleton desing pattern"""
+
+    _instances = {}
+
+    @synchronised(lock)
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
+
+
+class StocksInterface(metaclass=Singleton):
+    def get_last_close_price(self, market_id):
+        raise NotImplementedError("Must implement get_last_close_price")