+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added tooltips to UI widgets
 - Icon in status bar that shows internet connection status
 - Show application version in About dialog
+- Support for yfinance module to fetch stocks data
 
 ## [2.1.1] - 2020-01-13
 ### Changed